requirements.txt
pandas==2.1.4
mysql-connector-python==8.2.0
streamlit==1.29.0
plotly==5.18.0
gitpython==3.1.40
python-dotenv==1.0.0
numpy==1.26.2

